refactor: UserDetails 구현 클래스 수정
- 구현 클래스명 변경
- 구현 클래스 내 password 추가
