--- src/main/java/com/std/tothebook/security/UserDto.java
@@ -1,60 +1,0 @@
-package com.std.tothebook.security;
-
-import lombok.Builder;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.Collection;
-
-public class UserDto implements UserDetails {
-
-    private long id;
-
-    private String email;
-
-    private String nickname;
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return null;
-    }
-
-    @Override
-    public String getPassword() {
-        return null;
-    }
-
-    @Override
-    public String getUsername() {
-        return email;
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-
-    protected UserDto() {}
-
-    @Builder
-    public UserDto(long id, String email, String nickname) {
-        this.id = id;
-        this.email = email;
-        this.nickname = nickname;
-    }
-}
